Optimized matrix multiplication

+++ net_decimal.hpp
@@ -16,8 +16,7 @@ net_set<uint8_t> dec_polynomial_add(bool &ans_sgn, const net_set<uint8_t> &coe_a
              carry   += conj_a * coe_a[i] + conj_b * coe_b[i];
         auto curr_bit = carry % 10;
              carry   /= 10;
-        if(curr_bit < 0)
-        {
+        if(curr_bit < 0) {
             curr_bit = 10 + curr_bit;
             --carry;
         }
@@ -44,7 +43,7 @@ net_set<uint64_t> dec_polynomial_mult(const net_set<uint8_t> &coe_a, const net_s
     {
         auto curr_a = coe_a[i],
              curr_b = coe_b[j];
-        ans[i + j] += (uint64_t)curr_a * curr_b;
+        ans[i + j] += curr_a * curr_b;
     }
     return ans;
 }
